requirements.txt
Flask==2.3.3
langchain==0.1.20
langchain-community==0.2.10
langchain-google-genai==1.0.10
faiss-cpu==1.7.4
motor==3.3.2
google-generativeai==0.7.2
python-dotenv==1.0.0
PyPDF2==3.0.1
pypdf==3.17.4
pymongo==4.6.2
tiktoken==0.5.2
numpy==1.24.4
aiofiles==23.2.1